+++ src/defaultOptions.js
@@ -1,32 +1,56 @@
-/*
-    NOTE:
-
-    1. Block comments identify the start of options for that preset/plugin
-    2. Custom options have their description in the line comment above
-    3. Options that have a line comment next to them show their real key(path)
-    4. Options for external presets/plugins are described on their respective sites
-*/
-
 export default {
 
-    /* babel-preset-futagozaryuu */
+    /*********************** babel-preset-futagozaryuu ***********************/
 
-    // if true, will not use `babel-plugin-module-resolver`
+    /**
+     * If true, will not use `babel-plugin-module-resolver`.
+     * 
+     * @default false
+     * @type {boolean}
+     */
     "disableResolver": false,
 
-    // if true, will use a `babelrc.shouldPrintComment` method to ignore eslint comments
+    /**
+     * If true, will use the method `babelrc.shouldPrintComment` to ignore eslint comments.
+     * 
+     * @default true
+     * @type {boolean}
+     */
     "removeEslintComments": true,
 
-    // if true, will not exclude `transform-regenerator`
+    /**
+     * By default this preset will exclude `@babel/transform-regenerator`, but if this
+     * option is enabled, it will not perform the exclusion.
+     * 
+     * _NOTE:_ This does not mean that `@babel/transform-regenerator` will be included.
+     * 
+     * @default false
+     * @type {boolean}
+     */
     "regenerator": false,
 
-    // if true or an object, will include the plugin `transform-runtime`
+    /**
+     * These values will enable the plugin `@babel/transform-runtime` and:
+     * 
+     * 1. On `true` set the configuration I usually use with this plugin.
+     * 2. On a string, will also set the module name of the runtime that will be used.
+     * 3. On a object, will overwrite my configuration based on the object.
+     * 
+     * @default false
+     * @type {boolean|string|{}}
+     */
     "runtime": false,
 
-    // stages 0 - 3 will enable the appropriate `@babel/preset-stage-*`, where as 4 (or more) will disable them.
+    /**
+     * Any value that is `0` to `3` will enable the appropriate `@babel/preset-stage-*` preset,
+     * where as any other value will not add any preset's for proposal based plugins.
+     * 
+     * @default 0
+     * @type {number}
+     */
     "stage": 0,
 
-    /* @babel/preset-env */
+    /*********************** @babel/preset-env ***********************/
     // https://www.npmjs.com/package/@babel/preset-env#options
 
     "browsers": false, // targets.browsers
@@ -40,7 +64,7 @@ export default {
     "useBuiltIns": "usage",
     "forceAllTransforms": false,
 
-    /* babel-plugin-module-resolver */
+    /*********************** babel-plugin-module-resolver ***********************/
     // https://www.npmjs.com/package/babel-plugin-module-resolver#options
 
     "alias": {},
@@ -51,11 +75,18 @@ export default {
     "resolver": void 0, // resolvePath
     "root": "src",
 
-    /* @babel/plugin-transform-async-to-generator */
+    /*********************** @babel/plugin-transform-async-to-generator ***********************/
     // https://www.npmjs.com/package/@babel/plugin-transform-async-to-generator#usage
 
-    // if `true` will include `@babel/plugin-transform-async-to-generator`
-    // if an object, it will pass on to the plugin as options
+    /**
+     * These values enable the plugin `@babel/plugin-transform-async-to-generator` and:
+     * 
+     * 1. On `true` will not pass any options.
+     * 2. On a object, will overwrite my configuration based on the object.
+     * 
+     * @default "bluebird.coroutine"
+     * @type {boolean|{}}
+     */
     "async": {
 
         "module": "bluebird",