+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * DEV: Set minimum Node.js to v4.2
 * Added warning for minimum Node.js change
+* Renamed core `babel-*` dependencies to `@babel/*`
+* Updated `@babel/*` dependencies to `7.0.0-beta.31`
 
 <a name="2.11.0"></a>
 ## [v2.11.0](https://github.com/futagoza/babel-preset-futagozaryuu/compare/v2.10.0...v2.11.0) (2017-10-19)