+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added warning for minimum Node.js change
 * Renamed core `babel-*` dependencies to `@babel/*`
 * Updated `@babel/*` dependencies to `7.0.0-beta.31`
+* DEV: Change JSDoc type parameter for `object` to `{}` (why did I use it anyway?)
 
 <a name="2.11.0"></a>
 ## [v2.11.0](https://github.com/futagoza/babel-preset-futagozaryuu/compare/v2.10.0...v2.11.0) (2017-10-19)