+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * DEV: Set minimum Node.js to v4.2
 * Added warning for minimum Node.js change
+* Updated comment's for the default options (VS Code can now find them)
 * Renamed core `babel-*` dependencies to `@babel/*`
 * Updated `@babel/*` dependencies to `7.0.0-beta.31`
 * DEV: Change JSDoc type parameter for `object` to `{}` (why did I use it anyway?)