+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * DEV: Change JSDoc type parameter for `object` to `{}` (why did I use it anyway?)
 * Updated options for transforming `async` function's
 * Replaced references to core `babel-*` dependencies with `@babel/*`
+* Added support for the `shippedProposals` option passed to `@babel/preset-env`
 
 <a name="2.11.0"></a>
 ## [v2.11.0](https://github.com/futagoza/babel-preset-futagozaryuu/compare/v2.10.0...v2.11.0) (2017-10-19)