+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Temporary fix for babel/babel#6759 based on babel/babel#6745
 * DEV: Added ignore file for ESLint (Required for VSCode/ESLint extension to work properly)
 * Disabled loose mode (temporary fix for babel/babel#6805)
+* DEV: Removed `preversion` script (run before `npm run version`)
 
 <a name="2.11.0"></a>
 ## [v2.11.0](https://github.com/futagoza/babel-preset-futagozaryuu/compare/v2.10.0...v2.11.0) (2017-10-19)